django_websocket_chatting/requirements.txt
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
Django==4.1.12
django-bootstrap5==23.3
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.7
packaging==23.2
pillow==10.3.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
sentry-sdk==1.32.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
zope.interface==6.1
